% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_experiments.R
\name{colony_variance_experiment}
\alias{colony_variance_experiment}
\title{Experiment 2: Ant Colony Variance Decomposition}
\usage{
colony_variance_experiment(
  n_ants_values = c(10, 20, 30, 50, 100),
  p_explore_values = seq(0, 1, by = 0.2),
  n_replicates = 50,
  n_sites = 5,
  site_qualities = c(10, 8, 6, 4, 2)
)
}
\arguments{
\item{n_ants_values}{Integer vector of colony sizes}

\item{p_explore_values}{Numeric vector of exploration probabilities}

\item{n_replicates}{Monte Carlo replicates}

\item{n_sites}{Number of candidate nest sites}

\item{site_qualities}{True site quality vector}
}
\value{
A data.frame with columns \code{n_ants, p_explore, accuracy_mean,
  accuracy_sd, var_mean, cor_mean}
}
\description{
Sweeps colony size and exploration probability, computing decision
accuracy and the empirical variance/correlation of ant assessments.
}
