% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_visualization.R
\name{create_isomorphism_schematic}
\alias{create_isomorphism_schematic}
\title{Figure 1: Isomorphism Schematic}
\usage{
create_isomorphism_schematic(rf_results, colony_results)
}
\arguments{
\item{rf_results}{Output of \code{\link{variance_decomposition_experiment}}}

\item{colony_results}{Output of \code{\link{colony_variance_experiment}}}
}
\value{
A ggplot object
}
\description{
Plots theoretical \eqn{\rho = \rho_{max}(1-\theta)} alongside empirical
values from both random forest and ant colony experiments.
}
