% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_visualization.R
\name{plot_colony_accuracy}
\alias{plot_colony_accuracy}
\title{Supplementary: Colony Accuracy vs Size}
\usage{
plot_colony_accuracy(colony_results)
}
\arguments{
\item{colony_results}{Output of \code{\link{colony_variance_experiment}}}
}
\value{
A ggplot object
}
\description{
Supplementary: Colony Accuracy vs Size
}
