% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2_visualization.R
\name{plot_convergence_boost}
\alias{plot_convergence_boost}
\title{Plot Figure 4: Convergence Rates}
\usage{
plot_convergence_boost(results)
}
\arguments{
\item{results}{Output of \code{convergence_experiment_boost()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Figure 4: Convergence Rates
}
