% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_visualization.R
\name{plot_correlation_decay}
\alias{plot_correlation_decay}
\title{Figure 3: Correlation Decay Comparison}
\usage{
plot_correlation_decay(iso_results)
}
\arguments{
\item{iso_results}{Output of \code{\link{isomorphism_test}}}
}
\value{
A ggplot object
}
\description{
Figure 3: Correlation Decay Comparison
}
