% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2_visualization.R
\name{plot_margin_quorum}
\alias{plot_margin_quorum}
\title{Plot Figure 3: Margin vs Quorum}
\usage{
plot_margin_quorum(boost_result, X, y, acar_result)
}
\arguments{
\item{boost_result}{Output of \code{adaboost()}.}

\item{X}{Predictor matrix used in boosting.}

\item{y}{Label vector used in boosting.}

\item{acar_result}{Output of \code{acar()}.}
}
\value{
A combined \code{patchwork} plot.
}
\description{
Plot Figure 3: Margin vs Quorum
}
