% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2_visualization.R
\name{plot_noise_robustness_boost}
\alias{plot_noise_robustness_boost}
\title{Plot Figure 5: Noise Robustness}
\usage{
plot_noise_robustness_boost(results)
}
\arguments{
\item{results}{Output of \code{noise_experiment_boost()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Figure 5: Noise Robustness
}
