% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_visualization.R
\name{plot_optimal_decorrelation}
\alias{plot_optimal_decorrelation}
\title{Figure 4: Optimal Decorrelation}
\usage{
plot_optimal_decorrelation(optimal_results)
}
\arguments{
\item{optimal_results}{Output of \code{\link{optimal_decorrelation_experiment}}}
}
\value{
A ggplot object
}
\description{
Figure 4: Optimal Decorrelation
}
