% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part1_visualization.R
\name{plot_sensitivity_heatmap}
\alias{plot_sensitivity_heatmap}
\title{Figure 5: Sensitivity Heat-map}
\usage{
plot_sensitivity_heatmap(sensitivity_results)
}
\arguments{
\item{sensitivity_results}{Output of \code{\link{sensitivity_analysis}}}
}
\value{
A ggplot object
}
\description{
Figure 5: Sensitivity Heat-map
}
