% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2_visualization.R
\name{plot_weak_learnability}
\alias{plot_weak_learnability}
\title{Plot Figure 1: Weak Learnability Theorem}
\usage{
plot_weak_learnability(results)
}
\arguments{
\item{results}{Output of \code{weak_learnability_experiment()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Figure 1: Weak Learnability Theorem
}
