% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2_visualization.R
\name{plot_weight_pheromone}
\alias{plot_weight_pheromone}
\title{Plot Figure 2: Weight vs Pheromone Evolution}
\usage{
plot_weight_pheromone(boost_result, acar_result, site_qualities)
}
\arguments{
\item{boost_result}{Output of \code{adaboost()}.}

\item{acar_result}{Output of \code{acar()}.}

\item{site_qualities}{Numeric vector used in the ACAR call.}
}
\value{
A combined \code{patchwork} plot.
}
\description{
Plot Figure 2: Weight vs Pheromone Evolution
}
