% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2_adaboost.R
\name{predict_stump}
\alias{predict_stump}
\title{Predict with a Decision Stump}
\usage{
predict_stump(stump, X)
}
\arguments{
\item{stump}{A stump list from \code{find_best_stump}.}

\item{X}{Numeric predictor matrix.}
}
\value{
Numeric prediction vector in \{-1, +1\}.
}
\description{
Predict with a Decision Stump
}
