% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part2_adaboost.R
\name{track_weights}
\alias{track_weights}
\title{Track AdaBoost Weight Evolution}
\usage{
track_weights(X, y, M = 50)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Label vector.}

\item{M}{Number of iterations.}
}
\value{
An M x n matrix of instance weights.
}
\description{
Convenience wrapper that returns only the weight history.
}
